pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "crab-scholar"
version = "0.2.0"
authors = [
  { name="imnotdev25" },
]
description = "Research paper analysis pipeline with citation crawling, pluggable LLM prompts, and knowledge graph building"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["research", "papers", "knowledge-graph", "llm", "citation-analysis", "evaluation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.1",
    "httpx>=0.27.0",
    "networkx>=3.2",
    "litellm>=1.0.0",
    "kreuzberg>=4.0.0",
    "beautifulsoup4>=4.12.0",
    "unidecode>=1.3.0",
    "pyvis>=0.3.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.9.0",
    "respx>=0.21.0",
]

[project.scripts]
crab = "crab_scholar.cli:app"

[project.urls]
Repository = "https://github.com/imnotdev25/crab-scholar"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.hatch.build.targets.wheel]
packages = ["src/crab_scholar"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.uv.sources]
crabscholar = { workspace = true }
